Move tests to __tests__ folders
