+++ src/ts/shared/errors/__tests__/DataValidationError.test.ts
@@ -1,4 +1,4 @@
-import DataValidationError from './DataValidationError';
+import DataValidationError from '../DataValidationError';
 
 const validationErrors = { foo: 'foo' };
