+++ src/ts/shared/errors/__tests__/InvalidParameterError.test.ts
@@ -1,4 +1,4 @@
-import InvalidParameterError from './InvalidParameterError';
+import InvalidParameterError from '../InvalidParameterError';
 
 const parameterName = 'parameterName';
