+++ src/ts/shared/errors/__tests__/RecordVersionError.test.ts
@@ -1,17 +1,17 @@
-import RecordLockedError from './RecordLockedError';
+import RecordVersionError from '../RecordVersionError';
 
 test('should instantiate error with expected instance properties', (): void => {
-	const error = new RecordLockedError();
+	const error = new RecordVersionError();
 	const expected = {
-		name: 'RecordLockedError',
-		message: 'Record is locked and cannot be updated',
+		name: 'RecordVersionError',
+		message: 'Record has changed since it was read and cannot be updated',
 	};
 	expect(error).toMatchObject(expected);
 });
 
 test('should allow for override of message', (): void => {
 	const message = 'foo';
-	const error = new RecordLockedError({
+	const error = new RecordVersionError({
 		message,
 	});
 	expect(error.message).toEqual(message);
